Refactor pet-listing to petListing in file names
